Move Command and CommandLineParser into mzutils
