+++ src/main.cpp
@@ -3,18 +3,23 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "commandlineparser.h"
+
+#include <QCoreApplication>
+
 #include "leakdetector.h"
 #include "platforms/windows/windowsutils.h"
 #include "stdio.h"
-#ifdef MZ_WINDOWS
+#include "version.h"
 
+#ifdef MZ_WINDOWS
 #  include <windows.h>
 
 #  include <iostream>
 
 #  include "platforms/windows/windowsutils.h"
-
 #endif
+
+constexpr const char* CLP_DEFAULT_COMMAND = "ui";
 
 Q_DECL_EXPORT int main(int argc, char* argv[]) {
 #ifdef MZ_DEBUG
@@ -35,6 +40,10 @@ Q_DECL_EXPORT int main(int argc, char* argv[]) {
   WindowsUtils::lockDownDLLSearchPath();
 #endif
 
+  QCoreApplication::setApplicationName("Mozilla VPN");
+  QCoreApplication::setOrganizationName("Mozilla");
+  QCoreApplication::setApplicationVersion(APP_VERSION);
+
   CommandLineParser clp;
-  return clp.parse(argc, argv);
+  return clp.parse(argc, argv, CLP_DEFAULT_COMMAND);
 }
