+++ src/utils/tests/CMakeLists.txt
@@ -4,6 +4,7 @@
 
 # The tests
 qt_add_executable(utest-chacha20poly testchacha20poly.cpp testchacha20poly.h)
+qt_add_executable(utest-commandlineparser testcommandlineparser.cpp testcommandlineparser.h)
 qt_add_executable(utest-curve25519 testcurve25519.cpp testcurve25519.h)
 qt_add_executable(utest-hkdf testhkdf.cpp testhkdf.h)
 qt_add_executable(utest-ipaddress testipaddress.cpp testipaddress.h)
