+++ src/utils/tests/testcommandlineparser.cpp
@@ -4,6 +4,8 @@
 
 #include "testcommandlineparser.h"
 
+#include <QtTest/QtTest>
+
 #include "command.h"
 #include "commandlineparser.h"
 
@@ -65,12 +67,10 @@ void TestCommandLineParser::basic() {
   QFETCH(int, result);
 
   CommandLineParser clp;
-  QCOMPARE(clp.parse(argc, argv), result);
+  QCOMPARE(clp.parse(argc, argv, "ui"), result);
   QCOMPARE(clp.argc(), argc);
   QCOMPARE(clp.argv(), argv);
 
   QFETCH(bool, uiExecuted);
   QCOMPARE(s_executed, uiExecuted);
 }
-
-static TestCommandLineParser s_testCommandLineParser;
