+++ src/utils/tests/testcommandlineparser.h
@@ -2,9 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "helper.h"
+#include <QObject>
 
-class TestCommandLineParser : public TestHelper {
+#include "testhelper.h"
+
+class TestCommandLineParser : public QObject, TestHelper<TestCommandLineParser> {
   Q_OBJECT
 
  private slots:
