+++ tests/unit/CMakeLists.txt
@@ -43,10 +43,6 @@ target_sources(unit_tests PRIVATE
     ${MZ_SOURCE_DIR}/captiveportal/captiveportal.h
     ${MZ_SOURCE_DIR}/connectionhealth.cpp
     ${MZ_SOURCE_DIR}/connectionhealth.h
-    ${MZ_SOURCE_DIR}/command.cpp
-    ${MZ_SOURCE_DIR}/command.h
-    ${MZ_SOURCE_DIR}/commandlineparser.cpp
-    ${MZ_SOURCE_DIR}/commandlineparser.h
     ${MZ_SOURCE_DIR}/constants.cpp
     ${MZ_SOURCE_DIR}/constants.h
     ${MZ_SOURCE_DIR}/controller.h
@@ -276,8 +272,6 @@ target_sources(unit_tests PRIVATE
     testconnectionhealth.h
     testcontroller_p.cpp
     testcontroller_p.h
-    testcommandlineparser.cpp
-    testcommandlineparser.h
     testcryptosettings.cpp
     testcryptosettings.h
     testdnshelper.cpp
